add condition on adding branch_id

+++ database/migrations/tenant/2025_11_27_051118_add_column_branch_id_to_cashier_sessions.php
@@ -12,12 +12,14 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('cashier_sessions', function (Blueprint $table) {
-            $table->unsignedBigInteger('branch_id')->nullable()->after('cashier_id');
+            if (!Schema::hasColumn('cashier_sessions', 'branch_id')) {
+                $table->unsignedBigInteger('branch_id')->nullable()->after('cashier_id');
 
-            $table->foreign('branch_id')
-                ->references('id')
-                ->on('branches')
-                ->onDelete('set null');
+                $table->foreign('branch_id')
+                    ->references('id')
+                    ->on('branches')
+                    ->onDelete('set null');
+            }
         });
     }
 
@@ -27,8 +29,10 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('cashier_sessions', function (Blueprint $table) {
-            $table->dropForeign(['branch_id']);
-            $table->dropColumn('branch_id');
+            if (Schema::hasColumn('cashier_sessions', 'branch_id')) {
+                $table->dropForeign(['branch_id']);
+                $table->dropColumn('branch_id');
+            }
         });
     }
 };
